Core: GetIniKey improvements
git-svn-id: svn://ultimatepp.org/upp/trunk@1647 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Stream.cpp
@@ -850,7 +850,7 @@ void StringStream::Reserve(int n)
 {
 	SetWriteMode();
 	intptr_t p = ptr - buffer;
-	wdata.SetLength(GetSize() + n);
+	wdata.SetLength((int)GetSize() + n);
 	SetWriteBuffer();
 	ptr = buffer + p;
 }
